refactored TestExecutor
