+++ source/DasBlog.SmokeTest/Tester.cs
@@ -18,10 +18,13 @@ namespace DasBlog.SmokeTest
 		private readonly IBrowser browser;
 		public TestResults Results { get; } = new TestResults();
 		private readonly Pages pages;
-		public Tester(IBrowser browser)
+		private ITestExecutor testExecutor;
+
+		public Tester(IBrowser browser, ITestExecutor testExecutor)
 		{
 			this.browser = browser;
 			pages = new Pages(browser);
+			this.testExecutor = testExecutor;
 		}
 
 		public void Test()
@@ -41,66 +44,13 @@ namespace DasBlog.SmokeTest
 				, new TestStep(() => pages.Login.Password.Text.ToLower().Contains("the password field is required"))
 				, new TestStep(() => pages.Login.IsDisplayed())
 			};
-			ExecuteTestSteps(testSteps);
+			testExecutor.Execute(testSteps, Results);
 		}
 
-		private void ExecuteTestSteps(List<TestStep> testSteps, [CallerMemberName]string testName = null)
-		{
-			foreach (var step in testSteps)
-			{
-				switch (step.Value)
-				{
-					case Action action:
-						action();
-						break;
-					case Func<bool> func:
-						bool result = func();
-						if (!result)
-						{
-							Results.Add(testName, false, (step.Description as Expression<Func<bool>>).ToString());
-							return;
-						}
-						break;
-					default:
-						throw new Exception("Unknow test step encountered");
-				}
-			}
-			Results.Add(testName, false, string.Empty);
-		}
 
 		public void Dispose()
 		{
 			browser?.Dispose();
-		}
-	}
-
-
-	public class TestStep
-	{
-		private Expression<Action> action;
-		private Expression<Func<bool>> func;
-		public TestStep(Expression<Action> action)
-		{
-			System.Diagnostics.Debug.Assert(action != null);
-			this.action = action;
-		}
-
-		public TestStep(Expression<Func<bool>> func)
-		{
-			System.Diagnostics.Debug.Assert(func != null);
-			this.func = func;
-		}
-		public object Value
-		{
-			get
-			{
-				return action == null ? (object)func?.Compile() : (object)action?.Compile();
-			}
-		}
-
-		public object Description
-		{
-			get { return action == null ? (object) func : (object) action; }
 		}
 	}
 }
